requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
pydantic==2.4.2
mysql-connector-python==8.2.0
requests==2.31.0
schedule==1.2.0
concurrent-log-handler==0.9.24
python-multipart==0.0.6